backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.26.1
pytesseract==0.3.10
Pillow==10.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
